Refactoring of benchmarks: Separation of meta-data from training code
